+++ firebase-segmentation/src/androidTest/java/com/google/firebase/segmentation/local/CustomInstallationIdCacheTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.segmentation;
+package com.google.firebase.segmentation.local;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertNull;
@@ -28,7 +28,9 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-/** Instrumented tests for {@link CustomInstallationIdCache} */
+/**
+ * Instrumented tests for {@link com.google.firebase.segmentation.local.CustomInstallationIdCache}
+ */
 @RunWith(AndroidJUnit4.class)
 public class CustomInstallationIdCacheTest {
