+++ firebase-segmentation/src/main/java/com/google/firebase/segmentation/local/CustomInstallationIdCache.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.segmentation;
+package com.google.firebase.segmentation.local;
 
 import android.content.Context;
 import android.content.SharedPreferences;
